+++ .github/workflows/ci.yml
@@ -26,7 +26,7 @@ jobs:
           cache-on-failure: true
 
       - name: cargo test
-        run: cargo test --all
+        run: cargo test --all --all-features
 
   lint:
     runs-on: ubuntu-latest
