+++ bins/revme/src/statetest/mod.rs
@@ -2,7 +2,6 @@ mod cmd;
 pub mod merkle_trie;
 pub mod models;
 mod runner;
-mod trace;
 
 pub use cmd::Cmd;
 pub use runner::TestError as Error;
