+++ crates/revm/Cargo.toml
@@ -15,7 +15,7 @@ auto_impl = { version = "1.0", default-features = false }
 bytes = { version = "1.1", default-features = false }
 futures = { version = "0.3.24", optional = true }
 hashbrown = { version = "0.13" }
-hex = { version = "0.4", optional = true }
+hex = { version = "0.4", default-features = false }
 num_enum = { version = "0.5", default-features = false } # used for SpecId from u8 cast
 revm_precompiles = { path = "../revm_precompiles", version = "1.1.2", default-features = false }
 rlp = { version = "0.5", default-features = false } # used for create2 address calculation
@@ -51,9 +51,10 @@ no_gas_measuring = []
 optional_block_gas_limit = []
 optional_eip3607 = []
 optional_gas_refund = []
-std = ["bytes/std", "num_enum/std", "rlp/std"]
+std = ["bytes/std", "num_enum/std", "rlp/std", "hex/std"]
 secp256k1 = ["revm_precompiles/secp256k1"]
 k256 = ["revm_precompiles/k256_ecrecover"]
+ethersdb = ["tokio", "futures", "ethers-providers", "ethers-core"]
+with-serde = ["serde", "hex/serde", "hashbrown/serde", "ruint/serde"]
+# deprecated feature
 web3db = []
-ethersdb = ["tokio", "futures", "ethers-providers", "ethers-core", "hex"]
-with-serde = ["serde", "hex", "hex/serde", "hashbrown/serde"]
