+++ crates/revm/src/lib.rs
@@ -14,7 +14,9 @@ mod journaled_state;
 mod models;
 mod specification;
 
+// Reexport primary types.
 pub use bits::{B160, B256};
+pub use bytes::Bytes;
 pub use ruint::aliases::U256;
 
 pub use evm_impl::{create2_address, create_address, EVMData, Host};
@@ -24,7 +26,6 @@ pub type DummyStateDB = InMemoryDB;
 pub use db::{Database, DatabaseCommit, InMemoryDB};
 pub use evm::{evm_inner, new, EVM};
 pub use gas::Gas;
-pub use inspector::{GasInspector, Inspector, NoOpInspector};
 pub use instructions::{
     opcode::{self, spec_opcode_gas, OpCode, OPCODE_JUMPMAP},
     Return,
@@ -40,7 +41,11 @@ extern crate alloc;
 
 pub(crate) const USE_GAS: bool = !cfg!(feature = "no_gas_measuring");
 
-// reexport `revm_precompiles`
+/// reexport `revm_precompiles`
 pub mod precompiles {
     pub use revm_precompiles::*;
 }
+
+/// Reexport Inspector implementations
+pub use inspector::inspectors;
+pub use inspector::Inspector;
